Update README.
The primary change is to support more RFCs than I originally planned
after reviewing them further and seeing how easy they are to implement
I'm also not sure how extensible I can or want to make this library, so
I removed that goal. I'll re-evaluate that later when I get a better
idea how I'm going to structure the client and/or server.